newsboard/requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.3.0
black==20.8b1
celery==4.4.7
click==7.1.2
dj-database-url==0.5.0
Django==3.1.1
django-celery-beat==2.0.0
django-timezone-field==4.0
djangorestframework==3.11.1
flake8==3.8.3
kombu==4.6.11
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
vine==1.3.0
whitenoise==5.2.0
gunicorn==19.6.0